Simple Neural Network
This is an implementation of a two-layer neural network.The training method is stochastic gradient descent with momentum. It computes XOR for the given input. It uses two activation functions, one for each layer. One is a tanh function and the other is the sigmoid function. It uses [cross-entropy] as it's loss function.